IAM | User Inline Policy Document Additional Validation
